+++ src/TestableHttpClient/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: CLSCompliant(true)]
-[assembly: InternalsVisibleTo("TestableHttpClient.NFluent")]
 [assembly: InternalsVisibleTo("TestableHttpClient.Tests")]